refactor: Move metadata iterator to Iterator directory
